+++ lib/scss_lint/linter/nesting_depth.rb
@@ -3,6 +3,8 @@ module SCSSLint
   class Linter::NestingDepth < Linter
     include LinterRegistry
 
+    IGNORED_SELECTORS = [Sass::Selector::Parent, Sass::Selector::Pseudo]
+
     def visit_root(_node)
       @max_depth = config['max_depth']
       @depth = 1
@@ -10,8 +12,11 @@ module SCSSLint
     end
 
     def visit_rule(node)
-      if @depth > @max_depth
-        add_lint(node, "Nesting should be no greater than #{@max_depth}, but was #{@depth}")
+      if ignore_selectors?(node)
+        yield
+      elsif @depth > @max_depth
+        add_lint node, "Nesting should be no greater than #{@max_depth}, " \
+                       "but was #{@depth}"
       else
         # Only continue if we didn't exceed the max depth already (this makes
         # the lint less noisy)
@@ -19,6 +24,20 @@ module SCSSLint
         yield # Continue linting children
         @depth -= 1
       end
+    end
+
+  private
+
+    def ignore_selectors?(node)
+      simple_selectors(node.parsed_rules).all? do |selector|
+        IGNORED_SELECTORS.include? selector.class
+      end if config['ignore_parent_selectors']
+    end
+
+    def simple_selectors(node)
+      node.members.flat_map(&:members).reject do |simple_sequence|
+        simple_sequence.is_a? String
+      end.flat_map(&:members)
     end
   end
 end