+++ CHANGELOG.md
@@ -15,6 +15,8 @@
   by the `SelectorFormat` linter
 * Add `TrailingWhitespace` linter which reports whitespace characters at the
   end of a line
+* Change `NestingDepth` to allow parent selectors to be ignored in depth count
+  by passing configuration option `ignore_parent_selectors: true`
 
 ## 0.39.0